+++ app/src/client/enums/videoPalettes.js
@@ -9,6 +9,8 @@ export const FCEU_15 = 'fceu-15';
 export const FCEUX = 'fceux';
 export const NESTOPIA_RGB = 'nestopia-rgb';
 export const NESTOPIA_YUV = 'nestopia-yuv';
+export const SONY_CXA2025AS = 'sony-cxa2025as';
+export const UNSATURATED_V6 = 'unsaturated-v6';
 
 export const values = [
   ASQ_REAL_A,
@@ -20,6 +22,8 @@ export const values = [
   FCEUX,
   NESTOPIA_RGB,
   NESTOPIA_YUV,
+  SONY_CXA2025AS,
+  UNSATURATED_V6,
 ];
 
 export const labels = {
@@ -32,6 +36,8 @@ export const labels = {
   [FCEUX]: 'FCEUX',
   [NESTOPIA_RGB]: 'Nestopia (RGB)',
   [NESTOPIA_YUV]: 'Nestopia (YUV)',
+  [SONY_CXA2025AS]: 'SONY CXA2025AS US',
+  [UNSATURATED_V6]: 'Unsaturated V6',
 };
 
 export const options = createOptions(values, labels);